Refactor: Auth-Logik in services/auth_service gebuendelt
Passwort-Hashing, Session-Token-Verwaltung, Rollen-Hierarchie,
valid_voice_parts und das user_response-DTO lagen als private Helfer in
api/auth.py und wurden aus admin.py, guest_links.py, seed.py sowie
services/guest_link_service.py und services/folder_types.py per
Cross-Router-Import gezogen. Die Logik lebt jetzt in
backend/services/auth_service.py; api/auth.py ist ein duenner Wrapper um
Login/Register/Logout-Endpoints.

Tests auf feature/test-infra muessen bei Merge angepasst werden
(backend/tests/conftest.py + test_guest_links.py importieren noch die
alten Namen aus api.auth — auf main sind keine Tests betroffen).
